requirements.txt
pillow
pyyaml
